Add RTREE_NOATOMICS option
Add the compiler option RTREE_NOATOMIC to disable using an
atomic reference counter. Instead, a normal int will be used
in its place.

This means that using rtree_clone() will no longer result in a
clone that is thread-safe, but it will still be a valid clone for
single threaded programs. This may be desired (or needed) for
environments where there are no atomics, or where concurrency is
coroutine-based.

Usage:
    cc -DRTREE_NOATOMICS rtree.c